@taldcroft 2024-08-21
My 2c is that we should ignore PLR1730 (https://docs.astral.sh/ruff/rules/if-stmt-min-max/):
- As @mhvk noted it can be slower.
- There are situations where the intent of the original code is more clear to me, e.g.
```
if count < 0:
    count = 0
# vs
count = max(count, 0)
```

We already ran into this PLR1730 in our Chandra code and it just ended up causing discussions that are distracting and waste time.