@taldcroft 2024-08-21
See https://github.com/astropy/astropy/issues/16863#issuecomment-2302519231