@pllim 2024-08-21
Benchmark failures seem unrelated? 🤷‍♀️ 

```
| Change   | Before [81432542]    | After [4440822a]    |   Ratio | Benchmark (Parameter)                               |
|----------|----------------------|---------------------|---------|-----------------------------------------------------|
| +        | 45.4±0.1μs           | 70.0±30μs           |    1.54 | units.time_quantity_times_unit                      |
| +        | 247±20μs             | 322±6μs             |    1.31 | units.TimeQuantityOpLargeArray.time_quantity_square |

SOME BENCHMARKS HAVE CHANGED SIGNIFICANTLY.
PERFORMANCE DECREASED.
```