@taldcroft 2024-08-26
> So I see 2 approvals but also links to comments that this slows things down. Should we still merge?

These changes are not within tight loops so potential sub-microsec changes will have no noticeable impact. Also, one of the cases removes a redundant `len()` call so the execution is not necessarily slower.